# crates/fleet-api/src/openapi.yaml
openapi: 3.0.3
info:
  title: Grok Fleet Orchestrator — HTTP API
  description: >-
    Worker registration/heartbeat/join, bootstrap-token management, worker
    credentials, host inventory registration, and Prometheus metrics for
    `fleet serve`'s `/v1` HTTP API.


    This spec covers only the `/v1` and `/metrics` surface served by the
    `fleet-api` crate. The separate dashboard HTTP API (`fleet-dashboard`,
    `--dashboard-bind`, cookie-session + RBAC authenticated, ~30 more routes
    under `/api/*`) is out of scope — see
    `docs/architecture/api-reference.md` for a partial reference.


    Hand-maintained from `crates/fleet-api/src/{schema.rs,handlers.rs,app.rs}`.
    If it drifts from the code, the code is the source of truth — please
    file a correction.
  version: "0.1.0"
  license:
    name: (see repository LICENSE)
...
servers:
  - url: /v1
    description: Default mount point (bind address configured via `--http-bind`)
security:
  - bearerAuth: []
...
paths:
...
  /workers/register:
    post:
      operationId: registerWorker
      tags: [workers]
      summary: Register or re-register a worker
      description: >-
        Used when the orchestrator already knows the worker's endpoint
        (e.g. `fleet provision`). For self-service join with a bootstrap
        token, use `POST /workers/join` instead. When `--transport acp` is
        active, registration immediately attempts to open the ACP WebSocket
        session; a connection failure still updates the Store but leaves the
        worker undispatchable until HealthChecker demotes it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /workers/join:
    post:
      operationId: joinWorker
      tags: [workers, bootstrap-tokens]
      summary: Self-service worker join using a bootstrap token
      description: >-
        Consumes a bootstrap token atomically (single `UPDATE ... RETURNING`,
        so two concurrent submissions of the same single-use token cannot
        both succeed) and renders a complete `worker.toml` for the caller to
        write to disk.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRequest"
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinResponse"
        "401":
          description: Token not found, expired, or exhausted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: An active worker with the same `name` already exists at a different endpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workers/heartbeat:
    post:
      operationId: heartbeatWorker
      tags: [workers]
      summary: Periodic worker status update
      description: >-
        `grok_version`/`fleet_worker_version`/`os_info` are not just
        recorded — every heartbeat upserts the corresponding `hosts`
        inventory row. Omitting them means the host inventory view goes
        stale for that worker.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeartbeatRequest"
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HeartbeatResponse"
        "404":
          description: Unknown worker_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workers:
    get:
      operationId: listWorkers
      tags: [workers]
      summary: List workers
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/WorkerStatus"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
        - name: label_*
          in: query
          description: >-
            Label filter, one query param per label key, e.g.
            `?label_arch=arm64&label_gpu=true`. Not a literal parameter
            named `label_*` — OpenAPI has no first-class syntax for a
            dynamic-key query param family, so this row documents the
            convention instead.
          schema:
            type: string
      responses:
        "200":
          description: Worker summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkerSummary"
  /workers/{id}:
    get:
      operationId: getWorker
      tags: [workers]
      summary: Get a single worker
      parameters:
        - $ref: "#/components/parameters/WorkerId"
      responses:
        "200":
          description: Worker detail (same shape as the list items)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerSummary"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deregisterWorker
      tags: [workers]
      summary: Deregister a worker
      description: >-
        Calls transport.unregister (tearing down the ACP WebSocket session,
        if any) before deleting the Store row.
      parameters:
        - $ref: "#/components/parameters/WorkerId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeregisterRequest"
      responses:
        "200":
          description: >-
            Deregistered. Note: 200 with a JSON body, not 204 No Content.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeregisterResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /workers/{name}/credentials:
    put:
      operationId: putWorkerCredential
      tags: [credentials]
      summary: Set or rotate a worker's LLM API credential
      description: >-
        Sets exactly one (model_id, api_key) pair per call. `api_key` is
        sent in plaintext over this authenticated channel and encrypted at
        rest (AES-256-GCM via `FLEET_MASTER_KEY`). Requires the master key
        to be configured — returns 503 otherwise.
      parameters:
        - $ref: "#/components/parameters/WorkerName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutCredentialRequest"
      responses:
        "200":
          description: Stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutCredentialResponse"
        "503":
          description: Master key not configured — credentials API disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listWorkerCredentials
      tags: [credentials]
      summary: List a worker's stored credentials (metadata only)
      description: Never returns the decrypted `api_key` — see the `/export` endpoint for that.
      parameters:
        - $ref: "#/components/parameters/WorkerName"
      responses:
        "200":
          description: Credential summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CredentialSummary"
  /workers/{name}/credentials/{model_id}/export:
    get:
      operationId: exportWorkerCredential
      tags: [credentials]
      summary: Export a decrypted credential (provisioning use only)
      description: >-
        Returns the decrypted `api_key` plus a rendered `grok_config_section`
        TOML snippet for the worker's `~/.grok/config.toml`. Requires bearer
        auth like every other `/v1` route — there is no additional
        elevated-permission check.
      parameters:
        - $ref: "#/components/parameters/WorkerName"
        - $ref: "#/components/parameters/ModelId"
      responses:
        "200":
          description: Decrypted credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportedCredential"
        "404":
          $ref: "#/components/responses/NotFound"
  /workers/{name}/credentials/{model_id}:
    delete:
      operationId: deleteWorkerCredential
      tags: [credentials]
      summary: Remove a stored credential
      parameters:
        - $ref: "#/components/parameters/WorkerName"
        - $ref: "#/components/parameters/ModelId"
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted:
                    type: boolean
                    description: false if no matching credential existed
  /bootstrap-tokens:
    post:
      operationId: createBootstrapToken
      tags: [bootstrap-tokens]
      summary: Issue a new bootstrap token
      description: >-
        Despite the name, there is no admin/operator role distinction in
        fleet-api — any caller holding a valid bearer token for this server
        may call this endpoint.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBootstrapTokenRequest"
      responses:
        "201":
          description: >-
            The full token string is returned exactly once here. Store it in a
            secret manager; subsequent list and revoke calls use token_id and
            never reveal or accept the raw token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateBootstrapTokenResponse"
        "400":
          description: bytes out of [8, 256] range, or max_uses < 1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listBootstrapTokens
      tags: [bootstrap-tokens]
      summary: List all bootstrap tokens
      description: >-
        No `status` or `created_by` field is returned — clients must compute
        usability themselves from `use_count`/`max_uses`/`expires_at`.
      responses:
        "200":
          description: Token summaries, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BootstrapTokenSummary"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        `Authorization: Bearer <token>` against the `--api-tokens`/
        `FLEET_API_TOKENS` allow-list (constant-time compare). If
        `--cf-audience` is set, `CF-Access-Jwt-Assertion` is checked first
        and satisfies auth on its own — this spec doesn't model that
        alternate scheme. If neither is configured, the server runs in
        no-auth mode (dev only) and this requirement is not enforced.
...
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    WorkerStatus:
      type: string
      enum: [online, degraded, offline, circuit_open]
...
    RegisterResponse:
      type: object
      required: [worker_id, heartbeat_interval_secs, config_revision, orchestrator_version, status]
      properties:
        worker_id:
          type: string
          format: uuid
        heartbeat_interval_secs:
          type: integer
        config_revision:
          type: integer
          description: Currently always 1 — not a real revision counter yet
        orchestrator_version:
          type: string
        status:
          type: string
          enum: [registered]
...